+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Interval.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE PolyKinds #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 {- | This module is meant to be imported qualified, as some of its identifiers
@@ -18,10 +19,12 @@ module Plutarch.LedgerApi.Interval (
   pto,
   palways,
   pinterval,
+  pintervalOpenEnd,
 
   -- ** Queries
   pmember,
   pcontains,
+  ptryContains,
   pbefore,
   pafter,
 
@@ -275,6 +278,23 @@ pmember = phoistAcyclic $ plam $ \x i -> pcontains # i # (psingleton # x)
 
 {- | @'pcontains' # x # y@ is true if the interval @y@ is entirely contained in @a@.
 
+= Important note
+
+The only intervals supported by Plutus are one of the following:
+
+* @[start, +inf)@
+* @(-inf, end]@
+* @[start, end)@
+* @(-inf, +inf)@
+
+This function is written under the assumption that _only_ such intervals will
+be passed to it, and /will/ give wrong answers if not. Checking the
+invariants involved is somewhat expensive, and thus, this function remains
+for cases where efficiency is paramount.
+
+If you are unsure whether the interval you have is of a valid form, use
+'ptryContains', as this will error instead of giving you a wrong answer.
+
 @since 2.1.1
 -}
 pcontains ::
@@ -295,6 +315,21 @@ pcontains = phoistAcyclic $
     let lowerY = getField @"from" y
     let upperY = getField @"to" y
     pure $ leqP # (lToE # lowerX) # (lToE # lowerY) #&& leqP # (uToE # upperY) # (uToE # upperX)
+
+{- | As 'pcontains', but verifies that the intervals passed as arguments are
+sensible (see 'pcontains' description for this). Will error if given
+intervals that aren't supported.
+
+@since 3.2.2
+-}
+ptryContains ::
+  forall (a :: S -> Type) (s :: S).
+  (POrd a, PIsData a) =>
+  Term s (PInterval a :--> PInterval a :--> PBool)
+ptryContains = phoistAcyclic $ plam $ \x' y' -> unTermCont $ do
+  pcheckInterval x'
+  pcheckInterval y'
+  pure $ pcontains # x' # y'
 
 {- | Given @x@, create the interval @[x, x]@.
 
@@ -438,6 +473,7 @@ pafter = phoistAcyclic $
 
 @since 2.1.1
 -}
+{-# DEPRECATED pinterval "Use pintervalOpenEnd" #-}
 pinterval ::
   forall (a :: S -> Type) (s :: S).
   Term
@@ -451,6 +487,23 @@ pinterval = phoistAcyclic $
     let start = pcon $ PFinite $ pdcons @"_0" # x # pdnil
         end = pcon $ PFinite $ pdcons @"_0" # y # pdnil
      in pclosedInterval # start # end
+
+{- | As 'pinterval', except the endpoint is open.
+
+@since 3.2.2
+-}
+pintervalOpenEnd ::
+  forall (a :: S -> Type) (s :: S).
+  Term s (PAsData a :--> PAsData a :--> PInterval a)
+pintervalOpenEnd = phoistAcyclic $
+  plam $ \x y ->
+    let start = pcon $ PFinite $ pdcons @"_0" # x # pdnil
+        end = pcon $ PFinite $ pdcons @"_0" # y # pdnil
+        startClosure = pconstantData True
+        endClosure = pconstantData False
+        lower = pcon . PLowerBound $ pdcons @"_0" # pdata start #$ pdcons @"_1" # startClosure # pdnil
+        upper = pcon . PUpperBound $ pdcons @"_0" # pdata end #$ pdcons @"_1" # endClosure # pdnil
+     in pcon . PInterval $ pdcons @"from" # pdata lower #$ pdcons @"to" # pdata upper # pdnil
 
 -- Helpers
 
@@ -721,3 +774,67 @@ gtE' x = \case
   PFinite r ->
     let y = pfield @"_0" # r
      in y #< x
+
+-- Helpers
+
+-- Checks that we have one of the following:
+--
+
+-- * `[start, +inf)`
+
+-- * `(-inf, end]`
+
+-- * `[start, end)`
+
+-- * `(-inf, +inf)`
+pcheckInterval ::
+  forall (a :: S -> Type) (s :: S).
+  Term s (PInterval a) ->
+  TermCont s ()
+pcheckInterval i' = do
+  i <- tcont $ pletFields @'["from", "to"] i'
+  start <- tcont $ pletFields @'["_0", "_1"] $ getField @"from" i
+  end <- tcont $ pletFields @'["_0", "_1"] $ getField @"to" i
+  isCase1 <- tcont $ \k ->
+    pif
+      (getField @"_1" start)
+      -- Case 1: start is closed.
+      -- Check that it's finite, yield PTrue
+      ( pmatch (getField @"_0" start) $ \case
+          PFinite _ -> k (pcon PTrue)
+          _ -> ptraceInfoError "interval error: closed start not finite"
+      )
+      -- Case 2: start is open.
+      -- Check that it's -inf, yield PFalse
+      ( pmatch (getField @"_0" start) $ \case
+          PNegInf _ -> k (pcon PFalse)
+          _ -> ptraceInfoError "interval error: open start not -inf"
+      )
+  tcont $ \k ->
+    pif
+      isCase1
+      -- End must be open, and either finite or +inf
+      ( pif
+          (getField @"_1" end)
+          -- End is closed, fail
+          (ptraceInfoError "interval error: closed start and end")
+          -- End is open, check for form
+          ( pmatch (getField @"_0" end) $ \case
+              PNegInf _ -> ptraceInfoError "interval error: -inf end"
+              _ -> k ()
+          )
+      )
+      -- End must be either closed and finite, or open and +inf
+      ( pif
+          (getField @"_1" end)
+          -- End is closed, check if it's finite
+          ( pmatch (getField @"_0" end) $ \case
+              PFinite _ -> k ()
+              _ -> ptraceInfoError "interval error: closed end not finite"
+          )
+          -- End is open, check if it's +inf
+          ( pmatch (getField @"_0" end) $ \case
+              PPosInf _ -> k ()
+              _ -> ptraceInfoError "interval error: open end not +inf"
+          )
+      )