+++ plutarch-orphanage/src/PlutusLedgerApi/V1/Orphans/Interval.hs
@@ -2,16 +2,14 @@
 
 module PlutusLedgerApi.V1.Orphans.Interval () where
 
-import Data.Word (Word32)
 import PlutusLedgerApi.V1 qualified as PLA
-import PlutusLedgerApi.V1.Interval qualified as Interval
 import PlutusLedgerApi.V1.Orphans.Time ()
 import Test.QuickCheck (
   Arbitrary (arbitrary, shrink),
   Arbitrary1 (liftArbitrary, liftShrink),
   CoArbitrary (coarbitrary),
   Function (function),
-  frequency,
+  elements,
   functionMap,
   oneof,
   variant,
@@ -146,90 +144,48 @@ instance Function a => Function (PLA.UpperBound a) where
 {- | We provide an instance specialized to 'PLA.POSIXTime', rather than a more
 general one, as it doesn't make much sense to talk about 'PLA.Interval's of
 arbitrary types in general. Furthermore, this is the only instance we
-actually use, so there's no real loss there.
+actually use.
 
-This instance tries to make time intervals as fairly as possible, while also
-ensuring that they're sensibly formed. We work under the assumption of a
-32-bit epoch: while this is _technically_ not going to last much longer,
-we're safe until about 2030 on that basis, which should be enough for now.
+This instance aims to fit the following conditions, /in that order/:
 
-We choose not to shrink intervals, as this is surprisingly complex: in at
+1. Ensure that we construct permissible intervals (see
+https://github.com/Plutonomicon/cardano-transaction-lib/issues/1041#issuecomment-1280868255);
+and
+2. The intervals are generated as fairly as possible.
+
+We choose not to shrink intervals, as this is surprisingly complex. In at
 least one common case, it's not even possible to write a shrinker that will
-ever 'bottom out', due to us having infinite bounds!
+ever \'bottom out\', due to infinite bounds being a thing.
 
 @since 1.0.0
 -}
 instance Arbitrary (PLA.Interval PLA.POSIXTime) where
   {-# INLINEABLE arbitrary #-}
   arbitrary = do
-    let epochSize = fromIntegral (maxBound :: Word32)
-    lowerBound <-
-      frequency
-        [ (1, pure PLA.NegInf)
-        , (1, pure PLA.PosInf)
-        , (epochSize, PLA.Finite <$> arbitrary)
+    -- We have four options:
+    --
+    -- \* [start, +inf)
+    -- \* (-inf, end]
+    -- \* [start, end)
+    -- \* (-inf, +inf)
+    lo <-
+      oneof
+        [ PLA.LowerBound . PLA.Finite <$> arbitrary <*> pure True
+        , pure . PLA.LowerBound PLA.NegInf $ False
         ]
-    case lowerBound of
-      -- With a finite lower bound, it makes sense to talk about an upper one
-      PLA.Finite x -> do
-        lowerClosure <- arbitrary
-        let lower = PLA.LowerBound lowerBound lowerClosure
-        -- To ensure we generate something sensible for the upper bound, we
-        -- either generate a 'diff', or positive infinity.
-        whatUpper <-
-          frequency
-            [ (1, pure . Left $ PLA.PosInf)
-            , (epochSize, Right <$> arbitrary)
-            ]
-        case whatUpper of
-          -- If we have an infinite upper bound, we know it will be open.
-          Left _ -> do
-            let upper = PLA.UpperBound PLA.PosInf False
-            pure . PLA.Interval lower $ upper
-          Right diff -> case (diff, lowerClosure) of
-            -- A diff of 0 means we can only have a singleton closure sensibly.
-            (0, _) -> pure . Interval.singleton $ x
-            -- A diff of 1 with an open lower bound means we either have a
-            -- singleton closure or an empty one.
-            (1, False) -> do
-              upperClosure <- arbitrary
-              pure $
-                if upperClosure
-                  then Interval.singleton x
-                  else Interval.never
-            -- A diff of 1 with a closed lower bound is either a singleton
-            -- closure or one with two values.
-            (1, True) -> do
-              upperClosure <- arbitrary
-              pure $
-                if upperClosure
-                  then PLA.Interval lower . PLA.UpperBound (PLA.Finite (x + diff)) $ upperClosure
-                  else Interval.singleton x
-            -- A diff bigger than 1 can be treated uniformly.
-            (_, _) -> PLA.Interval lower . PLA.UpperBound (PLA.Finite (x + diff)) <$> arbitrary
-      -- With an negative infinite lower bound, we know it will be open.
-      PLA.NegInf -> do
-        let lower = PLA.LowerBound lowerBound False
-        -- To ensure we generate something sensible for the upper bound, we
-        -- do not attempt to generate NegInf
-        upperBound <-
-          frequency
-            [ (1, pure PLA.PosInf)
-            , (epochSize, PLA.Finite <$> arbitrary)
-            ]
-        case upperBound of
-          -- With a finite upper bound, we just choose a closure and move on.
-          PLA.Finite _ -> do
-            upper <- PLA.UpperBound upperBound <$> arbitrary
-            pure . PLA.Interval lower $ upper
-          -- With an infinite upper bound, we have the range that includes
-          -- everything. We use the canonical choice provided by
-          -- Interval.always.
-          _ -> pure Interval.always
-      -- With an positive infinite lower bound, we have the empty interval, and
-      -- can choose any representation of such that we like. We use the
-      -- canonical choice provided by Interval.never.
-      PLA.PosInf -> pure Interval.never
+    hi <- case lo of
+      PLA.LowerBound (PLA.Finite lo') _ -> do
+        delta <- arbitrary
+        elements
+          [ PLA.UpperBound (PLA.Finite (lo' + delta + 1)) False
+          , PLA.UpperBound PLA.PosInf False
+          ]
+      _ ->
+        oneof
+          [ PLA.UpperBound . PLA.Finite <$> arbitrary <*> pure True
+          , pure . PLA.UpperBound PLA.PosInf $ False
+          ]
+    pure . PLA.Interval lo $ hi
 
 -- | @since 1.0.0
 instance CoArbitrary a => CoArbitrary (PLA.Interval a) where